Read text from stdin
Whether passed via piping, like `cat foo.txt | werds` or pasted in when
`werds` is run with no file args and then the text is pasted in and
stdin is closed with Ctrl+D, it'll count the number of words.

This mirrors `wc` and seems useful for chaining together and using.

I tried to find an easy way to detect if anything was provided by
piping and otherwise handle it as if it no files were specified, but
that's not as simple as I was thinking, as opening up the stdin reader
is blocking. Oh well! I think this is fine for now.